fix: Update precision setting to 32 for full precision training

+++ config.py
@@ -71,7 +71,7 @@ class TrainingConfig:
 import torch
 import logging
 
-precision = 16  # Set to 16 for mixed precision, 32 for full precision, etc.
+precision = 32  # Set to 16 for mixed precision, 32 for full precision, etc.
 TRAIN_FROM_CHECKPOINT = False  # Set to True to resume training from a checkpoint
 # These parameters should be treated as modifiable from the main run_model.py script
